# backend/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Used only by /api/sonarr/webhook, since Sonarr/Radarr's built-in Webhook connection type has no custom-header support. Any username works; the password must be the API key.",
        "title": "Aura API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/validate/notifications": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send a test notification using the specified notification provider. This endpoint is used to verify that the notification settings are correct and that notifications can be sent successfully. The request should include the notification provider information, and the response will indicate whether the test notification was sent successfully or if there were any errors.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Validation"
                ],
                "summary": "Send Test Notification",
                "parameters": [
                    {
                        "description": "Notification Provider Information for sending the test notification",
                        "name": "provider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_validation.SendTestNotification_Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/routes_validation.SendTestNotification_Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized (only when Auth.Enabled=true)",
                        "schema": {
                            "$ref": "#/definitions/httpx.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.Config_CustomNotification": {
            "type": "object",
            "properties": {
                "enabled": {
                    "description": "Whether custom notifications are enabled.",
                    "type": "boolean"
                },
                "include_image": {
                    "description": "Whether to include an image with the custom notification.",
                    "type": "boolean"
                },
                "message": {
                    "description": "Message for the custom notification.",
                    "type": "string"
                },
                "title": {
                    "description": "Title for the custom notification.",
                    "type": "string"
                }
            }
        },
...
        "config.Config_Notification_Discord": {
            "type": "object",
            "properties": {
                "webhook": {
                    "description": "Webhook URL for the Discord notification provider.",
                    "type": "string"
                }
            }
        },
        "config.Config_Notification_Gotify": {
            "type": "object",
            "properties": {
                "api_token": {
                    "description": "Token for the Gotify notification provider.",
                    "type": "string"
                },
                "url": {
                    "description": "URL for the Gotify notification provider.",
                    "type": "string"
                }
            }
        },
        "config.Config_Notification_Provider": {
            "type": "object",
            "properties": {
                "discord": {
                    "description": "Discord notification settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.Config_Notification_Discord"
                        }
                    ]
                },
                "enabled": {
                    "description": "Whether this notification method is enabled",
                    "type": "boolean"
                },
                "gotify": {
                    "description": "Gotify notification settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.Config_Notification_Gotify"
                        }
                    ]
                },
                "provider": {
                    "description": "Notification provider",
                    "type": "string"
                },
                "pushover": {
                    "description": "Pushover notification settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.Config_Notification_Pushover"
                        }
                    ]
                },
                "webhook": {
                    "description": "Webhook notification settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.Config_Notification_Webhook"
                        }
                    ]
                }
            }
        },
        "config.Config_Notification_Pushover": {
            "type": "object",
            "properties": {
                "api_token": {
                    "description": "Token for the Pushover notification provider.",
                    "type": "string"
                },
                "user_key": {
                    "description": "UserKey for the Pushover notification provider.",
                    "type": "string"
                }
            }
        },
        "config.Config_Notification_Webhook": {
            "type": "object",
            "properties": {
                "headers": {
                    "description": "Headers for the Webhook notification provider.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "url": {
                    "description": "URL for the Webhook notification provider.",
                    "type": "string"
                }
            }
        },
...
        "httpx.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/logging.LogErrorInfo"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "httpx.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid or expired token"
                }
            }
        },
...
        "logging.LogErrorInfo": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "object"
                },
                "function": {
                    "type": "string",
                    "example": "Reload"
                },
                "help": {
                    "type": "string",
                    "example": "Check server logs for more details"
                },
                "line_number": {
                    "type": "integer",
                    "example": 42
                },
                "message": {
                    "type": "string",
                    "example": "Invalid or expired token"
                }
            }
        },
...
        "routes_validation.SendTestNotification_Request": {
            "type": "object",
            "properties": {
                "provider": {
                    "$ref": "#/definitions/config.Config_Notification_Provider"
                },
                "template": {
                    "$ref": "#/definitions/config.Config_CustomNotification"
                },
                "template_type": {
                    "type": "string"
                }
            }
        },
        "routes_validation.SendTestNotification_Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for programmatic/integration access. Generate one under Settings \u003e Authentication, then send it as this header on every request. This is the intended auth method for scripts and integrations - the session cookie above is for browser use only.",
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "SessionCookie": {
            "description": "Browser session cookie, set by POST /api/login or the OIDC callback. Not usable directly via Swagger \"Authorize\" - log in through the app UI in the same browser tab.",
            "type": "apiKey",
            "name": "aura_session",
            "in": "cookie"
        }
    }
}